+++ tests/xharness/XmlResultParser.cs
@@ -5,6 +5,7 @@ using System.Linq;
 using System.Xml;
 using System.Xml.Linq;
 using Xharness.Logging;
+using Xharness.Utilities;
 
 namespace Xharness {
